requirements.txt
tensorflow-cpu==2.16.1
keras>=3.0.0
protobuf==4.25.3
streamlit
joblib
Pillow
scikit-learn
numpy
pandas